+++ stacker/stack.py
@@ -87,12 +87,6 @@ class Stack(object):
 
     @property
     def requires(self):
-        # By definition, a locked stack has no dependencies, because we won't
-        # be performing an update operation on the stack. This means, resolving
-        # outputs from dependencies is unnecessary.
-        if self.locked and not self.force:
-            return []
-
         requires = set(self.definition.requires or [])
 
         # Add any dependencies based on output lookups